poc/requirements.txt
flask
rich
google-generativeai
python-dotenv
requests